Move urlUtils into a separate utils directory
