+++ lib/index.js
@@ -11,7 +11,7 @@ const configure = require('./configure')
 const resetConfiguration = require('./resetConfiguration')
 const emit = require('./emit')
 const invoke = require('./invoke')
-const urlUtils = require('./urlUtils')
+const urlUtils = require('./utils/url')
 
 const eventGateway = configuration => {
   if (
