+++ lib/listSubscriptions.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const urlUtils = require('./urlUtils')
+const urlUtils = require('./utils/url')
 
 module.exports = config =>
   config
